requirements.txt
watchdog
python-daemon
stuf
path.py
fabric